중복된 data class 하나로 통일
